refactor: Move DSL Text methods to DSL::Text
